Make BridgeStorageAccessor an interface and create BridgeStorageAccessorImpl to implement it. Move FeePerKbStorageProvider to /feeperkb package and make some methods protected instead of public.
