+++ rskj-core/src/test/java/co/rsk/test/builders/BridgeSupportBuilder.java
@@ -10,7 +10,8 @@ import co.rsk.peg.BtcBlockStoreWithCache.Factory;
 import co.rsk.peg.FederationSupport;
 import co.rsk.peg.btcLockSender.BtcLockSenderProvider;
 import co.rsk.peg.storage.BridgeStorageAccessor;
-import co.rsk.peg.storage.FeePerKbStorageProvider;
+import co.rsk.peg.storage.BridgeStorageAccessorImpl;
+import co.rsk.peg.feeperkb.FeePerKbStorageProvider;
 import co.rsk.peg.feeperkb.FeePerKbSupport;
 import co.rsk.peg.pegininstructions.PeginInstructionsProvider;
 import co.rsk.peg.utils.BridgeEventLogger;
@@ -93,7 +94,7 @@ public class BridgeSupportBuilder {
     }
 
     public BridgeSupport build() {
-        BridgeStorageAccessor bridgeStorageAccessor = new BridgeStorageAccessor(repository);
+        BridgeStorageAccessor bridgeStorageAccessor = new BridgeStorageAccessorImpl(repository);
 
         return new BridgeSupport(
             bridgeConstants,
